requirements.txt
kafka-python
requests
pyspark
streamlit
scikit-learn
joblib
plotly
streamlit-autorefresh
python-dotenv